requirement.txt
django==2.1.3
PyJWT==1.7.0
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
django-cors-headers==2.4.0
loguru==0.4.1
Pillow
timestring
django-sendgrid-v5